+++ src/App.jsx
@@ -1,8 +1,8 @@
 import React from "react";
-import Home from "./pages/Home/Home.jsx";
-import About from "./pages/About/About.jsx";
-import Events from "./pages/Events/Events.jsx";
-import Nav from "./Nav.jsx";
+import Home from "./pages/Home.jsx";
+import About from "./pages/About.jsx";
+import Events from "./pages/Events.jsx";
+import Nav from "./components/Nav.jsx";
 
 function App() {
   return (
