+++ src/components/Hero.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "../../../index.css";
+import "../index.css";
 import { motion } from "framer-motion";
 
 function Hero() {
