+++ src/pages/Home.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Hero from "./components/Hero";
+import Hero from "../components/Hero";
 
 function Home() {
   return <Hero />;
